symptom/requirements.txt
grpcio==1.27.2
grpcio-tools==1.27.2
protobuf==3.12.2

Jinja2==2.11.3
numpy>=1.17.4
tensorboardX==1.6
torch>=1.6.0
tqdm==4.55.1
transformers[sentencepiece]==4.1.1
scikit-learn==0.24.0
allennlp>=2.0.0
spacy>=3.0.0
markupsafe==2.0.1